store: split wizardModal slice into modular file structure
Splits the index.ts into slice.ts, selectors.ts, and a barrel index.ts
following the domain-split convention established by the blueprints,
cloudConfig, and env modularizations.
